break out createTemplate from init, add scripts base types
